Move backend_lookup method to a mixin.
